conflict file changed in b1

+++ app/src/main/java/com/farizma/truthdare/conflict.java
@@ -1,7 +1,7 @@
 package com.farizma.truthdare;
 
 public class conflict {
-    int roll;
+    float roll;
     String id,name;
 
     public conflict(int roll, String id, String name) {
